institutes seded

+++ database/migrations/2023_02_13_140445_create_institutes_table.php
@@ -27,7 +27,7 @@ return new class extends Migration
             $table->string('seo')->nullable();
             $table->boolean('active')->nullable()->default(false);
             $table->string('slug')->nullable();
-            $table->float('price');
+            $table->decimal('price', 8, 2);
             $table->timestamps();
         });
     }
